host-agent: make HostDaemon responsible for collecting snapshots
Having two periodic loops means that delays can be higher than
desired, especially when the state is collected just after updating
the cluster agent. With a 5s update frequency, the state being
reported would be 5s old and would stay at the cluster agent for
5s. So rather than have staleness bounded at 5s, we have 10s.

If we just collect the current state right before reporting, we
keep the delay at 5s. As a bonus, the logic becomes a little
simpler too.