alg: two fixes to HeypSigcomm20MaybeReviseLOPRIAdmission
First, we now do nothing when time doesn't flow forwards. This is
mainly useful for debugging tests, though it can happen if clocks
are out of sync.

Second, we now use the current LOPRI usage as the rate limit, not
the HIPRI+LOPRI usage - HIPRI limit. This is both simpler and makes
more sense since HIPRI usage really doesn't tell us much about how
much capacity there is available for LOPRI --- the LOPRI usage is
much better for that.