host-agent/enforcer-impl: add wrapper to invoke tc
Using TcCaller is both convenient and allows us to limit #includes
of boost headers to just tc-caller.cc. If we kept the tc-based
HostEnforcer and TcCaller all in one file, we'd have one file
that's really slow to compile and instead we get two moderately
slow files that can be compiled in parallel to one another.