app/testlopri: collect multiple types of latency
First, fix 'full' latency to be timed from when we intended the
RPC to be scheduled. This way we capture more delays at the client.
Second, collect latency from right when we are about to issue the
RPC. This is useful for measuring what is going on on the network
*right now*.