app/testlopri: err on running timers early
libuv only allows timers to be scheduled at the granularity of
milliseconds. So we have to decide when we schedule a timer,
should we round up or round down?

Previously, we rounded down meaning that the timer possibly fired
early. A recent commit changed this to round up. Turns out that
rounding up is not so great when you inter-arrival time is dozens
of *micro*seconds.

Now that we round down again, validation numbers are looking good.