cmd/flowselsim: parameterize along LOPRI frac and refactor output
The fraction of traffic that should use LOPRI is now a config
flag as it really is a part of the workload like the number of
hosts (and others). With this change, we can clearly see that the
greatest different between the intended split and the actual occurs
when the split = 0.5. (This happens because near 0 and 1, the error
in one direction is bounded - we can't be more than 0.1 below if we
only want to send 0.1 as LOPRI).

This commit also changes the output to use reflection as prep work
for dumping more metrics.