alg, cluster-agent: start adding new allocation policies
This adds the definitions for a few new policies and implements
one more downgrade policy (untested).

To implement DS_KNAPSACK_SOLVER, I am pulling in or-tools.
Unfortunately, or-tools, has its own fork of glog which will
conflict with the real glog if we link it in (and has collisions in
header files anyway). Sigh.

To "fix" this, switch all logging to use or-tools. If, in the
future, we need to change again, we can just change
heyp/log/logging.h rather than update all imports.