cmd/deploy-heyp: generate configs using starlark
Editing nearly-identical configs one by one is getting tedious and
is error prone. To fix this, the commit adds a starlark-based
config generator. The config script can lookup external addrs that
correspond to a particular cluster IP (no more updating configs
when we switch to a different cluster!) and will validate that the
generated protobuf messages are valid.

A follow-up commit will update experiments/web to use starlark.