fortio/stagedperiodic: fix scheduler
The scheduler was sending requests that were not spaced out
properly, fix this by removing the "tricky" line as our experiments
always execute more than a handful of requests.

The scheduler would also not handle multi-stage workloads correctly
because "start" was used as though it meant both the start of the
run and the start of the stage. Separate these into two separate
times.

Add a mockable clock implementation for testing. It has been
written so as to avoid making interface calls outside of testing.

Add tests for the scheduler.