experiments/web: reduce precision to 1s everywhere
I tried going back to mean usage over a time period (hence the
updated commented code), but it's very spikey, much more so than
what we see from the device counters and clients.

So abort, just use the EWMA counters. Even this is too spikey, so
drop precision to 1s which is a little bit better.

500ms is too spikey. The spikes don't seem real either given what