io: Communicate should read from the correct timeout fd
As it turns out, we had Communicate read from the same fd that
KillAfter was writing to. In fact, Communicate needed to read from
the fd that represented the other end of the pipe. This was all due
to the confusing flipping of fd indicies when we stored the result
of pipe().

We now keep the order of fds in timeout_pipe_ consistent with the
order returned by pipe().