host-agent: fix spikiness of usage measurements
The usage data we were collecting was very spikey even when we knew
the application was receiving data at a steady rate.

The problem was with how we collected usage data, namely, the
timestamp we were associating with the usage numbers was wrong.

We'd do this:

1. Collect usage counters
2. Assign timestamp
3. Determine whether usage was HIPRI/LOPRI

However, Step 3 required acquiring/releasing a lock in the host
enforcer once for *every flow*. Since the host enforcer can take a
while to change the system config, this can cause the timestamp
collected in Step 2 to be wildly incorrect, and we'd have the
spikiness result from that.

This commit changes this so we now acquire an uncontended lock once
before collecting usage counters and the timestamp, and we never
acquire it again.

With this, the usage data is looking very smooth.