deploy, proto, experiments/web: refactor for admission control
*** This commit breaks old fortio configs ***

Envoy performs admission control at the granularity of a listener.
This means that we should place backends into separate fortio
groups if we want them to be controlled separately.

Unfortunately, the multi-group mechanism was never fully fleshed
out: Envoy doesn't like to have multiple instances run on a machine
as this makes hot restart annoying (we can work around) and causes
resource interference between instances (hard to work around). In
particular, we are already high on CPU usage, so we really should
keep everything handled by one Envoy instance.

To do so, rearrange things so we put each fortio group into a
separate Envoy listener, and handle all listeners with the same
Envoy instance.

Verified that we currently see the same results as before. A
following change will enable admission control in the config.