cmd/proc-heyp, proc: add subcmd to diff remote configs
When debugging config issues, it's often helpful to see the
differences between the config files written to the cluster, as
that captures differences in deploy-heyp as well.

The implementation relies on git being present so that it can track
renames. Since this is just for debugging, this is probably OK.