host-agent: ignore job field in enforcement
We currently do not populate the job field when initializing the
simulated WAN. This is a problem because FlowAllocs from the
cluster agent do specify the job field. In practice, this leads to
some of the simulated WAN configs not applying.

The fix is to just ignore the job. It's unnecessary in the host
agent (everything on the host belongs to the same job).

Fixes #7