host-agent: create iptables rules to default all traffic to HIPRI
Previously, we would never create any iptables rules unless the
simulated WAN config instructed us to, or if the cluster agent told
us to. This led to the scenario where some traffic is not marked as
HIPRI and therefore sent at a lower priority level than intended.

Fix this by ensuring we have a rule in the SimulatedWanDB for all
pairs of DCs - if the user didn't specify one, we will use HIPRI
with no rate limit.