host-agent, cmd/deploy-heyp: collect fine-grained flow stats if requested
Previously we collected src dc, dst dc flow usage stats at the host.
To develop and tune a congestion detection mechanism, it would be
helpful to have full fine-grained statistics.

This commit adds all the plumbing we need to collect these stats.
A future commit will export additional metrics like the cwnd that
can be useful for developing and implementing the mechanism.