host-agent: implement complete parser for SS output
The current parser for SS output only collects the data we've been
interested in thus far: that is current usage and cumulative bytes
sent. For analysis purposes, we'd like to collect much more than
this, so this commit adds a parser for SS output that is much more
fully featured.

A future commit will replace SSFlowStateReporter's parser with this
one and plumb the auxillary state through.