deploy/actions: validate netem configs
In #11, we were seeing WA's approval be satisfied even though it
had a higher approval than AA and fewer hosts. It turns out that
the problem was we were, in fact, prioritizing WA because of QoS
bug.

Since 4151eab, host-agent sets outgoing traffic to use HIPRI at
startup. This was working for WA and we could see the iptables
rules being installed. For AA, however, we were setting netem_lopri
to an empty value. This is invalid, netem & netem_lopri must be
either null or have a set delay. This caused setup on AA hosts to
fail (though it's a non-fatal error) and thus, AA hosts would use
a lower QoS than HIPRI.

This commit validates the netem & netem_lopri configs when we
deploy so that this doesn't happen again. A following commit will
fix the problem in experiments/web/config.star.

Updates #11