internal/sampling, cmd/evalsampling: add ways to sample usage data
We're considering the possibility of sampling usage data to perform
QoS downgrade. We want to evaluate the accuracy of computing the
split using an estimate + using hashing to pick hosts to downgrade.

As a step towards this, this commit adds two ways to sample data:
- Uniform probabilty across hosts
- Uniform probabilty across bytes sent

In addition, this commit adds cmd/evalsampling to analyze the
effectiveness of these approaches.