alg, calg: bundle C++ algorithms for use in go
We will soon need to call the knapsack and max-min solvers from Go
for our intra-DC simulations. Since fast, quality solvers are
difficult to find, we'll use the ones we already have.

This commit adds a process by which we expose a C API for calling
the algorithms, a bazel target that collects that API and its
dependencies and forms a giant static library we can link from Go.
On Go's side, we add a new package, calg, that wraps these with a
cgo-free API.

Anything that depends on calg needs to set CGO_LDFLAGS to pass the
correct -L flag so that the linker can find libc++ and libc++abi.