intradc/montecarlo: flatten output for easier downstream use
Nested objects are easy enough to manipulate downstream (say in R),
but nested arrays are rather annoying. So this commit changes the
output so that we flatten arrays when generating the output.

This does make it more challenging to perform processing when we
want to compare systems on a per-instance basis. To compensate, we
add instance ids to the output, so that we can easily tell which
outputs correspond to the same instance. Additionally, all outputs
for the same instance are written next to each other in the output.