intradc/montecarlo: specialize estimateUsage based on sampler
After sorting, this function uses the largest chunk of runtime. Do
to the use of interfaces, function calls cannot be inlined and
everything goes on the heap.

Eliminate this overhead by switching on the type and calling a
version that has a concrete interface. A script generates both the
generic and specialized versions from a template.