intradc/montecarlo: add stats type and reflect-based summarization
The current method of copy-pasting a bunch of things is getting to
be unmanagable. Typos are hard to prevent and the same fields are
essentially being repeated (MeanV, VPerc, MeanAbsV, AbsVPerc).

The plan is to cut this down to just V and AbsV with all stats
under a single, new type. Additionally, this type will be populated
from the result types using the reflect-based populateSummary. So
there should be no data->summary typos any more.

This commit just adds the Stats type and populateSummary function.
A future commit will change everything to use them.