cluster-agent: use async callback API
Previously, we had the choice of using the easy and slow sync API,
or the tricky and fast async completion queue API. Now, grpc has a
better callback-driven API for C++, so switch to that.

This should give us better resource efficiency which will be
important as we evaluate the benefits of sampling. (Really this
helps baseline approaches, but we want a fair comparison).