cluster-agent: remove threads that are no longer needed
Previously, we used the sync grpc API. When the server wanted to
send data to a host, it would invoke the blocking API. But really,
we wanted to contact all hosts in parallel, so we sent data to each
host using separate threads.

Now that we use the async API, this should not be an issue. So the
threads are pure overhead. Remove them. (Once again, it is in our
interest that overhead is maximized, but we should optimize it to
have a fair comparison).