alg/downgrade: track state in DiffDowngradeSelectorImpl
When we try to respond quickly (say once every 500ms), it's
possible that we begin computing a new allocation before the
previous one has completed. As a result, the existing
DiffDowngradeSelectorImpl will forget about these WIP changes and
effectively undo them.

Fix by tracking the intended state of each child when we add diffs.
When we get around to implementing the more optimized controller
for downgrade (only), it will need to track state in a different
way anyway, so the performance is not too important.