bazel, host-agent: statically link libc into host-agent
host-agent will be used in a glibc-free environment (alpine linux
VMs), so we want to make sure that it is fully statically linked.

Unfortunately, a bug in older version of glibc (which, of course,
our dev servers happen to be running) causes problems when we pull
in a replacement malloc (tcmalloc). Fortunately, a workaround
exists so I added that to host-agent.cc.