host-agent/linux-enforcer: separately delete root qdisc on init
The root qdisc might just be the default and not be something we
can (or need to) delete. At the moment, this shouldn't cause any
problems because we ask tc to continue in case of error, but it
does cause an error message to be printed.

Silence the spurious message by running the delete+add individually
rather than as a batch and ignore any errors from the deletion.