virt/host: only masquerade ip for egress traffic
We want the VMs to see the original IP but not other nodes so that
the host agents that run in the VMs can attribute traffic to the
correct FG. This is OK because the VM still routes traffic out
through the host kernel, so when routing out the host will change
the source IP from the VM to itself.

With this, the cluster and host agents appear to be working.

TODO: adjust all the plotting stuff.