host-agent/linux-enforcer: fix a bunch of bunch with unthrottling
It turns out that there were a bunch of issue when we chose not
to rate limit HIPRI or LOPRI and there was buggy handling of rate
limits after first creating the simulated WAN. The latter issue
should be fairly harmless, but the former would result in many
unnecessary calls to tc.

This commit fixes all of that and adds some debugging facilities.

Test is upcoming.