host-agent/linux-enforcer: test LinuxHostEnforcer
We finally have the capability to test LinuxHostEnforcer, so do
that. The tests added in this commit found a number of performance
issues (in the form of unnecessary calls to tc and iptables), and
maybe one or two correctness issues.

These tests check that the enforcer makes both (1) the correct
number of calls to tc and iptables and (2) that the inputs to the
programs is what we expect.

Whew. Hopefully this fixes a bunch of problems.