fortio: use http2 (plaintext)
Turns out that this was a fairly straightforward change.
Reduces resource usage in envoy.

Also checked in some commented-out code to rate limit requests so
that we don't try to catch up when there is a backlog of requests.
Using it so far has made no difference or throttled too much.