intradc/feedbacksim: extract most of RunMultiIter into a state machine and test
RunMultiIter is not a trivial function, but also depends on enough
of ActiveScenario that we don't want to test it directly.

Instead, pull out the logic into a state machine and make
RunMultiIter a thin wrapper around that. Write the tests against
the state machine. This has found a few bugs.