intradc/feedbacksim: use a better estimator for coarseness
The number of flows is a bad estimator for coarseness as it does
not take skew into account. As an example, elephantsMice-10 can
have 0 usage on all except 10 tasks. Clearly the 1/number of tasks
is a bad estimator in this case.

An easy-to-track number that works is better is the max task usage.
Obviously we may not always have the max in the sample, but an
EWMA of the max seen usage seems to work fairly well.