cluster-agent: silence some clamping warnings
With feedback control, we add diffs to our LOPRI frac, and, as a
result, it's not unusual for the LOPRI frac to go out of the bounds
we expect.

Silence these warnings since nothing is wrong. In the feedforward
case, we don't expect the LOPRI frac to go out of bounds, so keep
the warning there.