alg: fix overage check and improve logging
Turns out we never checked the overage against the coarseness
condition - we were instead checking if the coarseness condition
was non-zero because C++ allows for implicit type conversions.

Fix the condition: feedback control now works!

Logging is also improved and can be toggled via a constant. Should
be fine to leave on for now.