all: initial dual-network config
The dual-net config offers two benefits: (1) the bottleneck is
between envoy and the backends and does not effect other traffic
and (2) we can run envoy on bigger machines. The second one is
really important as envoy appears to have been bottlenecked by CPU.

Not 100% there, but close. Filing TODOs on github.