host-agent/linux-enforcer: set a larger netem packet limit
Apparently, when we create a netem qdisc it has a limit for the
number of packets that will be queued. The size should be made
large enough for the BDP, at least.

We never set the limit explicitly, so we were getting netem's
default of 1000 packets. This is a very low limit and can cause us
to drop packets.

Bump the limit to 100,000. We could try to be smarter about it, but
that should be high enough for our experiments.