cmd/proc-heyp, proc: ignore unknown fields
We recently changed the cluster_agent_port scalar field to be a
cluster_agent_ports repeated field. This breaks processing existing
logs, but for no real reason: none of the analysis care about this
field.

Just ignore unknown fields from now on. We don't make many of these
sorts of changes, but we should be careful not to break anything in
the future.