host-agent/linux-enforcer: reduce calls to tc
Until now, we've been changing the host configuration in 3 steps:
first increase any rate limits, then flip QoS rules, and finish by
decreasing rate limits.

However, we only really care to change rate limits if that
particular HTB class is being used by some flow. If we see that no
flows are matched, we can safely ignore the change in rate limit.
This matters particularly with our current usage where each host is
given just 1 QoS, and when switching QoS we reducing the rate limit
for the unused QoS to 0.

This commit implements this optimization and performs better in
this case: we no longer change the rate limit for the unused class
and a combined change of QoS+limit is 1 call to tc + 1 call to
iptables. Changing just the rate limit or QoS is even better, we
only call one of the tools.